granular_qa_agent/pyproject.toml
[project]
name = "granular-qa-agent"
version = "0.1.0"
description = "Granular QA Agent - LLM-Based Data Validation"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv",
    "pandas>=2.0.0",
]

[project.scripts]
validate = "validator:main"

